requirements.txt
chromedriver_binary==2.43.0
requests==2.20.0
selenium==3.141.0
Pillow==5.2.0
